refactor(storage): Split implementations based on FLASH_BIT_ACCESS.
[no changelog]
