+++ core/embed/trezorhal/stm32f4/platform.h
@@ -23,7 +23,7 @@
 #include STM32_HAL_H
 #include <stdint.h>
 
-#define FLASH_BYTE_ACCESS 1
+#define FLASH_BIT_ACCESS 1
 
 typedef enum {
   CLOCK_180_MHZ = 0,
