+++ core/embed/trezorhal/unix/platform.h
@@ -1,4 +1,4 @@
 
-#define FLASH_BYTE_ACCESS 1
+#define FLASH_BIT_ACCESS 1
 
 void emulator_poll_events(void);
