+++ legacy/flash.h
@@ -24,7 +24,7 @@
 #include <stdlib.h>
 #include "secbool.h"
 
-#define FLASH_BYTE_ACCESS 1
+#define FLASH_BIT_ACCESS 1
 #define FLASH_SECTOR_COUNT 24
 
 #include "flash_common.h"
