+++ storage/tests/c/flash.h
@@ -25,7 +25,7 @@
 #include "secbool.h"
 
 #ifndef FLASH_QUADWORD
-#define FLASH_BYTE_ACCESS 1
+#define FLASH_BIT_ACCESS 1
 #endif
 
 #include "flash_common.h"
